api/requirements.txt
fastapi
uvicorn[standard]
azure-cosmos
pygeohash
python-dotenv
httpx
